updated with suspense wrapper changes
